added active to components

+++ dx11minimal/timer.cpp
@@ -1,7 +1,6 @@
 #include "timer.h"
 
 
-
 double timer::PCFreq = 0.0;
 __int64 timer::counterStart = 0;
 
@@ -10,10 +9,9 @@ double timer::frameEndTime = 0;
 double timer::nextFrameTime = 0;
 double timer::frameRenderingDuration = 0.0;
 
-DWORD timer::currentTime = 0;
-DWORD timer::lastFrameTime = 0;
-DWORD timer::deltaTime = 0;
-
+double timer::currentTime = 0;
+double timer::lastFrameTime = 0;
+double timer::deltaTime = 0;
 
 
 void timer::StartCounter()
